+++ avl_operations.h
@@ -4,6 +4,6 @@
 #include "avl_module.h"
 
 // retorna o fator de balanceamento da subárvore
-uint8_t balance_factor(node_t *nd);
+int balance_factor(node_t *nd);
 
 #endif
